Add spim syscalls for MIPS32
The list is based on James Larus' description of spim in the appendix of
Computer Organization and Design, by Patterson and Hennessy, on Figure
A.9.1 (page A-44). This appendix is freely available on Larus' homepage;
the table is on page 43 of http://pages.cs.wisc.edu/~larus/HP_AppA.pdf.